Adding a commented-out future change.
This would be the change to allow clicking directly on the world to move
(skipping the move button). Unfortunately, there's currently a problem
where clicking the buttons causes you to move, which is exacerbated by
this (since now clicking the _move_ button would cause you to move
immediately rather than setting the state). Therefore, for now the
change is commented out. We can uncomment it once we fix that other
problem.