comma-dangle: always-multiline
This means that multiline lists and objects always require a comma for
the last element, even if there's only one element. Ideally we'd have an
exception for that case, but I couldn't find a way, so it seems better
to enforce always (and we'll just deal with that case) than not to
enforce at all.