Add entities for the ground.
Clicking on the ground results in (attempting) to move to that square,
though this is only relevant if you _haven't_ yet clicked "move" since
otherwise the movement squares are on top of the ground and have the
same effect.

There aren't any walls yet, but we avoid creating MovementSquares on
squares that don't have a Ground, so you're effectively limited to the
part of the world that has a Ground.

Also a couple cleanups:
  - Fix that all GridObjects were SpaceFillingObjects. Now it's just
    Characters, so in particular we don't need a hack to stop
    MovementSquares from blocking movement.
  - Factor out isAdjacent into a new module for utility functions, along
    with the new getDistance.
  - Fix that the maximum movement distance was actually maxDistance - 1
    instead of maxDistance (inequalities shouldn't be strict), and
    factor that constant out into consts.js.