Partial refactor of components.
Mostly: introduce subclasses of GridObject for DynamicObject and
StaticObject. The idea is that everything will subclass those so we can
easily query for only scenery or non-scenery.

Also refactor the getPos/setPos/animateTo logic a bit while I was at it.

For now, I leave SpaceFillingObject, but it should probably be removed
in favor of an isPassable bit.