Enforce that all GridObjects have Mouse.
The only one that didn't have Mouse was Tree, and that was leading to a
bug where trying to move to a Tree gave the wrong error message ("Can't
move that far" rather than "Something's in the way"). Rather than just
add Mouse to Tree individually, let's fix the general problem.

Fixes #17.