Early beginnings of AI control.
This is a very rudimentary hack. For now, I have two functions which we
call in startCharacter, depending on whether a player or AI is moving.
The idea is that they would determine what move to make according to
different strategies (user input vs. some algorithm). For now, though,
the AI function just skips the turn because there isn't code to choose
its action.

More work will probably be needed when taking the action, since
different things need to be done after the action (see recent commit
that added callbacks to doMove doAttack etc.).