Mostly back out my previous change.
It can lead to an infinite loop when an AI-controlled unit is exactly
its speed away from a unit it considers an enemy. The AI generates a
move-and-attack action, which fails checkAction so doesn't happen. That
AI unit still has action points, so we generate another action for it.
Lather, rinse, repeat.