Skeleton file for new action classes.
This is the start of a branch for fixing all the long switch statements
based on ActionType (issue #35). Still need to actually move all that
logic into new_action.js (and probably give it a better name when we're
done).

The plan here is:
    1.  Get everything working in the new structure
    2.  Resolve cyclic imports
As such, for now I disable the linter rule against cyclic imports on
this branch. Don't merge this branch into master until that rule is
re-enabled!